Reorganization of template files
