+++ src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/helpers/aas-configuration-helpers.ts
