+++ src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/helpers/configuration-helper.ts
