+++ src/management-system-v2/app/(dashboard)/[environmentId]/machine-config/templates/parameter-template-organizationdata.ts
@@ -1,0 +1,14 @@
+import { defaultParameter } from '../helpers/configuration-helper';
+
+export function createOrgOrganizationDataParameter() {
+  const organizationParameter = defaultParameter(
+    'organization',
+    [{ text: 'Organization', language: 'en' }],
+    [],
+  );
+  const dataParameter = defaultParameter('data', [{ text: 'Data', language: 'en' }], []);
+
+  dataParameter.subParameters = [];
+  organizationParameter.subParameters = [dataParameter];
+  return organizationParameter;
+}
