+++ src/management-system-v2/components/aas-config-creation-button.tsx
@@ -9,7 +9,7 @@ import { useParams, useRouter } from 'next/navigation';
 import { useEnvironment } from './auth-can';
 import { useAddControlCallback } from '@/lib/controls-store';
 import { spaceURL } from '@/lib/utils';
-import { defaultTdsConfigurationTemplate } from '@/app/(dashboard)/[environmentId]/machine-config/configuration-templates-tds';
+import { defaultTdsConfigurationTemplate } from '@/app/(dashboard)/[environmentId]/machine-config/templates/configuration-templates-tds';
 type ConfigCreationButtonProps = ButtonProps & {
   customAction?: (values: { name: string; description: string }) => Promise<any>;
   wrapperElement?: ReactNode;
